Created README.md
It has been a while since I worked on this project.  I created it for a proof of concept that I was developing for my professional team, but once I got this in place, I had the foundation to take off with the real project, so this was laid aside for a season.